+++ crates/zeevonk/benches/data/gdtf_files/Generic@Dimmer@8_and_16bit_Generic_Dimmer.gdtf
