+++ crates/zeevonk/benches/data/gdtf_files/Generic@RGBW8@added_white_channel.gdtf
