+++ crates/zeevonk/benches/data/projects/realistic_single_universe/gdtf_files/Clay_Paky@Sharpy@ClayPaky_Official_File_Fw_Ver_2_25_006.gdtf
