+++ crates/zeevonk/benches/data/projects/realistic_single_universe/gdtf_files/Robe_Lighting%40Robin_600_LEDWash%402024-03-25__Zoom_range_fix.gdtf
